Remove `common`
In the file `lua_api.txt` there is no definition of which functions are common and which are ssm, so I decided to temporarily delete the directory common and place all the files by type, according to the documentation. I think that in this case it would be better to make symlinks so as not to disturb the order. Even if you decide to put everything back the way it was, it is better to do it when the entire documentation is complete.